backend-requirements.txt
absl-py==1.0.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.9b0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coverage==6.0.1
debugpy==1.5.1
decorator==5.1.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-rest-registration==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==0.8.2
Faker==9.3.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
IMDbPY==2021.4.18
importlib-metadata==4.10.1
ipykernel==6.9.0
ipython==8.0.1
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
libgravatar==1.0.0
lxml==4.8.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.27
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-surprise==1.1.1
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.1.4
surprise==0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0